dpaa_eth: add explicit cast where needed
Reviewed-on: http://git.am.freescale.net:8181/18839

+++ drivers/net/ethernet/freescale/dpa/dpaa_eth_common.c
@@ -1220,7 +1220,7 @@ int dpa_fq_init(struct dpa_fq *dpa_fq, bool td_enable)
 				dpa_fq->fq_type == FQ_TYPE_TX_CONF_MQ) {
 			initfq.we_mask |= QM_INITFQ_WE_CGID;
 			initfq.fqd.fq_ctrl |= QM_FQCTRL_CGE;
-			initfq.fqd.cgid = priv->cgr_data.cgr.cgrid;
+			initfq.fqd.cgid = (uint8_t)priv->cgr_data.cgr.cgrid;
 			/* Set a fixed overhead accounting, in an attempt to
 			 * reduce the impact of fixed-size skb shells and the
 			 * driver's needed headroom on system memory. This is
@@ -1268,7 +1268,7 @@ int dpa_fq_init(struct dpa_fq *dpa_fq, bool td_enable)
 				 dpa_fq->fq_type == FQ_TYPE_RX_PCD)) {
 			initfq.we_mask |= QM_INITFQ_WE_CGID;
 			initfq.fqd.fq_ctrl |= QM_FQCTRL_CGE;
-			initfq.fqd.cgid = priv->ingress_cgr.cgrid;
+			initfq.fqd.cgid = (uint8_t)priv->ingress_cgr.cgrid;
 			/* Set a fixed overhead accounting, just like for the
 			 * egress CGR.
 			 */
